Renamed build subcommand to convert
Also cleaned up some tests
